pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flowengine"
version = "0.1.0"
description = "Lightweight YAML-driven state machine for Python"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "FlowEngine Contributors" }]
keywords = [
    "workflow",
    "state-machine",
    "pipeline",
    "yaml",
    "orchestration",
    "automation",
    "component",
    "flow",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = ["pyyaml>=6.0", "pydantic>=2.0"]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "mypy>=1.8",
    "ruff>=0.2",
    "types-PyYAML>=6.0",
    "build>=1.0",
    "twine>=6.2.0",
    "httpx>=0.27",
]
http = ["httpx>=0.27"]

[project.urls]
Homepage = "https://github.com/yourorg/flowengine"
Documentation = "https://github.com/yourorg/flowengine#readme"
Repository = "https://github.com/yourorg/flowengine"
Changelog = "https://github.com/yourorg/flowengine/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.sdist]
include = ["/src"]

[tool.hatch.build.targets.wheel]
packages = ["src/flowengine"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["tests/"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.coverage.run]
source = ["src/flowengine"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
